backend/pyproject.toml
[project]
name = "production-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Minimal dependencies for Vercel deployment
    # Connects to external Qdrant - no local ML processing
    "fastapi>=0.124.4",
    "pydantic>=2.10.3",
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "google-generativeai>=0.8.0",
    # Removed heavy packages:
    # - inngest (workflow orchestration)
    # - llama-index-* (heavy ML packages)
    # - streamlit (web UI)
    # - uvicorn (provided by Vercel)
    # - sentence-transformers (800MB+)
]
